# src/openapi.yml
openapi: 3.0.0
info:
  title: Social-API
  version: 1.0.0
servers:
  - url: http://localhost:9001
  - url: https://anti-social-mongo-the-soldiers.onrender.com
paths:
...
  /api/comments:
    get:
      summary: Obtener todos los comentarios
      tags:
        - Comentarios
      responses:
        '200':
          description: Lista de comentarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '500':
          description: Error interno al obtener los comentarios

    post:
      summary: Crear un nuevo comentario
      tags:
        - Comentarios
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentInput'
      responses:
        '201':
          description: Comentario creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Error de validación al crear comentario
...
components:
  schemas:
...
    CommentInput:
      type: object
      required:
        - comment
        - userId
        - postId
      properties:
        comment:
          type: string
        creationDate:
          type: string
          format: date-time
        userId:
          type: string
          description: ID del usuario
        postId:
          type: string
          description: ID del post
        visible:
          type: boolean
    Comment:
      type: object
      properties:
        _id:
          type: string
        comment:
          type: string
        creationDate:
          type: string
          format: date-time
        userId:
          type: string
        postId:
          type: string
        visible:
          type: boolean
...